# project/backend/docs/openapi.yaml
openapi: 3.0.0
info:
  title: User & Game API
  version: 1.0.0
  description: API documentation for user management, authentication, and WebSocket routes.
servers:
  - url: https://localhost:3000
paths:
  /api/users:
    get:
      summary: Get all users
      description: Retrieve a list of all users in the database.
      operationId: getAllUsers
      responses:
        '200':
          description: A list of users.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    username:
                      type: string
        '401':
          description: Unauthorized. Invalid or missing token.
        '500':
          description: Internal Server Error.
    delete:
      summary: Delete a user
      description: Delete a user from the database.
      operationId: deleteUser
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - username
              properties:
                username:
                  type: string
      responses:
        '200':
          description: User successfully deleted.
        '403':
          description: Forbidden. Insufficient permissions to process the request.
        '422':
          description: Unprocessable Entity. Validation error.
        '500':
          description: Internal Server Error.
    post:
      summary: Create a new user
      description: Create a new user in the database.
      operationId: createUser
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - username
              properties:
                username:
                  type: string
      responses:
        '200':
          description: User created successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  username:
                    type: string
        '422':
          description: Unprocessable Entity. Validation error.
        '500':
          description: Internal Server Error.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT